# fair-gift-server/openapi.json
{"openapi":"3.0.0","paths":{"/login":{"post":{"operationId":"LoginController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Login"]}},"/login/logout":{"post":{"operationId":"LoginController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Login"]}},"/users":{"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/me":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/group/create":{"post":{"operationId":"GroupController_createGroup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"201":{"description":""}},"tags":["Group"]}},"/group/my-groups":{"get":{"operationId":"GroupController_getUserGroups","parameters":[],"responses":{"200":{"description":""}},"tags":["Group"]}},"/group/{groupId}/users/{userId}":{"post":{"operationId":"GroupController_addUserToGroup","parameters":[{"name":"groupId","required":true,"in":"path","schema":{"type":"number"}},{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddUserToGroupDto"}}}},"responses":{"201":{"description":""}},"tags":["Group"]}}},"info":{"title":"API Documentation","description":"Documentation API pour le système de gestion de groupes et utilisateurs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"JohnDoe@gmail.com","description":"Email de l'utilisateur"},"password":{"type":"string","example":"password123","description":"Mot de passe de l'utilisateur"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{}},"CreateGroupDto":{"type":"object","properties":{}},"AddUserToGroupDto":{"type":"object","properties":{}}}}}